Move libfabric initialization into protocols
We tried to share initialization (ie, finding the right provider
list) between the send/recv and rdma protocols, but this really
just made very complicated code that asked for a superset of the
required capabilities.  So split the initialization out into
the providers, with some of the common code moved into a new
ofiutils component.
