+++ include/nccl_ofi_sendrecv.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Amazon.com, Inc. or its affiliates. All rights reserved.
+ * Copyright (c) 2023-2024 Amazon.com, Inc. or its affiliates. All rights reserved.
  */
 
 #ifndef NCCL_OFI_SENDRECV_H_
@@ -221,9 +221,7 @@ typedef struct nccl_net_ofi_sendrecv_req {
 /*
  * @brief	Initialize plugin with sendrecv protocol structures
  */
-int nccl_net_ofi_sendrecv_init(struct fi_info *ofi_info_list,
-			       int num_ofi_infos,
-			       bool provide_own_mr_key,
+int nccl_net_ofi_sendrecv_init(const char *provider_filter,
 			       nccl_net_ofi_plugin_t **plugin_p);
 
 #ifdef _cplusplus
